backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.4.0
gunicorn==22.0.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Filtering and Search (CRITICAL - Missing!)
django-filter==23.5
# Geographic and GIS support (using basic calculations instead)

# Additional packages for enhanced features
boto3                    # AWS S3 integration for file storage
django-storages          # Django storage backends (S3, etc.)
Pillow                   # Image processing for profile pictures
django-phonenumber-field # Phone number validation and formatting
phonenumbers             # Phone number parsing and validation

# Phase 3.5: Global API Infrastructure
djangorestframework-throttling==0.1.0  # Rate limiting and throttling
django-redis==5.4.0                    # Redis caching backend
drf-spectacular==0.27.0                # OpenAPI/Swagger documentation
openpyxl==3.1.2                        # Excel export/import support
redis==5.0.1                           # Redis Python client
django-cacheops==8.1.0                 # Advanced caching operations
